[*] ApiError : added error handling and passing to client with ApiError

+++ AlgorithmsRanking/Controllers/AlgorithmsController.cs
@@ -5,6 +5,7 @@ using Microsoft.AspNetCore.Authorization;
 
 namespace AlgorithmsRanking.Controllers
 {
+    using AlgorithmsRanking.Models;
     using AlgorithmsRanking.Entities;
     using AlgorithmsRanking.Services;
 
@@ -23,50 +24,81 @@ namespace AlgorithmsRanking.Controllers
         [HttpGet("list")]
         public async Task<IActionResult> List()
         {
-            return Ok(await _db.GetAlgorithmsListItemsAsync());
+            var items = await _db.GetAlgorithmsListItemsAsync();
+
+            return Ok(items);
         }
 
         [HttpGet]
         public async Task<IActionResult> Get()
         {
-            return Ok(await _db.GetAlgorithmsAsync());
+            var items = await _db.GetAlgorithmsAsync();
+
+            return Ok(items);
         }
 
         [HttpGet("{id:int}")]
         public async Task<IActionResult> Get(int id)
         {
-            return Ok(await _db.GetAlgorithmAsync(id));
+            var model = await _db.GetAlgorithmAsync(id);
+
+            if (model == null)
+            {
+                return NotFound(new ApiError("404", "Not Found", $"Алгоритм #{id} не найден"));
+            }
+
+            return Ok(model);
         }
 
         [HttpPost]
         public async Task<IActionResult> Create([FromBody]Algorithm model)
         {
+            if (model == null)
+            {
+                return BadRequest(new ApiError("400", "Null model", $"{typeof(Algorithm)} cannot be null"));
+            }
+
             try
             {
                 return Ok(await _db.CreateAlgorithmAsync(model));
             }
             catch (Exception ex)
             {
-                return BadRequest(ex);
+                return BadRequest(new ApiError(ex));
             }
         }
         
         [HttpPut("{id:int}")]
         public async Task<IActionResult> Edit([FromRoute]int id, [FromBody]Algorithm model)
         {
+            if (model == null)
+            {
+                return BadRequest(new ApiError("400", "Null model", $"{typeof(Algorithm)} cannot be null"));
+            }
+
+            if ((await _db.GetAlgorithmAsync(id)) == null)
+            {
+                return NotFound(new ApiError("404", "Not Found", $"Алгоритм #{id} не найден"));
+            }
+
             try
             {
                 return Ok(await _db.UpdateAlgorithmAsync(id, model));
             }
             catch (Exception ex)
             {
-                return BadRequest(ex);
+                return BadRequest(new ApiError(ex));
             }
         }
         
         [HttpDelete("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute]int id)
         {
+            if ((await _db.GetAlgorithmAsync(id)) == null)
+            {
+                return NotFound(new ApiError("404", "Not Found", $"Алгоритм #{id} не найден"));
+            }
+
             try
             {
                 await _db.RemoveAlgorithmAsync(id);
@@ -75,7 +107,7 @@ namespace AlgorithmsRanking.Controllers
             }
             catch (Exception ex)
             {
-                return BadRequest(ex);
+                return BadRequest(new ApiError(ex));
             }
         }
     }
